Remove extra delete_entity call in GET GetAll tester
The extra call was causing one of the unit tests to fail; this
removal allows the test to pass.

In this case, handle_delete() returned code 500 (Internal server
error) for a table which existed, but an entity which no longer
existed in the table. My question is, is this expected behaviour
or should we be checking for it in handle_delete()?