Refactor well-formed request checks in handle_get BasicServer
This solves the problem of one of the tests returning an incorrect
result due to the check interpreting the request with the incorrect
operation.

In case you're wondering, the problem manifested some time ago
(at the beginning of this branch), and did not exist in the same
form before this commit.