Change CMake compilation for any directory structure
Previously, CMake was set up for a Vagrant virtual machine;
however, this was limiting, so the changes allow the project
to be set up in any environment without requiring a virtual
machine.

The UnitTest++ directory is now one directory up from before -
in the same directory as the Casablanca and Azure Storage
directories, for consistency.

I have also renamed the project in CMake as the project was
originally named BasicServer when it was first conceived, but has
grown beyond BasicServer.